Extract marker-waiting into ProjectUtils and only log on unexpected conditions
